# Conflicts:
#	src/main/resources/com/sd/mommyson/member/dao/MemberDAO.xml

+++ src/main/java/com/sd/mommyson/owner/controller/OwnerController.java
@@ -2,14 +2,18 @@ package com.sd.mommyson.owner.controller;
 
 import java.io.File;
 import java.io.IOException;
+import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.UUID;
 
 import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpSession;
+
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -18,11 +22,19 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.SessionAttributes;
+import org.springframework.web.bind.support.SessionStatus;
 import org.springframework.web.multipart.MultipartFile;
 
+import org.springframework.web.servlet.ModelAndView;
+import org.springframework.web.servlet.mvc.support.RedirectAttributes;
+
 import com.sd.mommyson.member.dto.MemberDTO;
+import com.sd.mommyson.member.service.MemberService;
 import com.sd.mommyson.owner.dto.CouponDTO;
+import com.sd.mommyson.owner.dto.ProductDTO;
+import com.sd.mommyson.owner.dto.TagDTO;
 import com.sd.mommyson.owner.service.OwnerService;
+import com.sd.mommyson.user.dto.ReviewDTO;
 
 @Controller
 @RequestMapping("/owner/*")
@@ -30,11 +42,15 @@ import com.sd.mommyson.owner.service.OwnerService;
 public class OwnerController {
 	
 	private OwnerService ownerService;
+	private BCryptPasswordEncoder passwordEncoder;
+	private MemberService memberService;
 	
 	@Autowired
-	public OwnerController(OwnerService ownerService) {
+	public OwnerController(OwnerService ownerService, MemberService memberService , BCryptPasswordEncoder passwordEncoder) {
 		
 		this.ownerService = ownerService;
+		this.memberService = memberService;
+		this.passwordEncoder = passwordEncoder;
 	}
 	
 	/* 사업자 마이페이지 메인화면 */
@@ -51,17 +67,44 @@ public class OwnerController {
 	}
 	
 	/* 쿠폰 발행 */
-	@RequestMapping("coupon")
-	public String coupon(@ModelAttribute("owner") MemberDTO member, Model model) {
+	@GetMapping("coupon")
+	public String coupon(@ModelAttribute("loginMember") MemberDTO member, Model model) {
 		
 		List<CouponDTO> coupon = ownerService.selectCoupon(member);
 		System.out.println(coupon);
 		
 		//					items 이름 , 리스트이름
 		model.addAttribute( "coupon" ,coupon);
-		
-		
+	
 		return "owner/coupon";
+	}
+	
+
+	
+	@PostMapping("coupon") 	  // couponDTO를 선언하면 자동으로 값이 담겨져 // memCode를 가져오려면 세션이 필요
+	public String couponInsert(@ModelAttribute CouponDTO coupon, RedirectAttributes ra, HttpSession session,  HttpServletRequest request) {
+															  // 리다이렉트를 해줄때 값을 넘겨주는...........
+		// @ModelAttribute 을 사용하는 순간 DTO에 필드값이랑 name값이 같으면 자동으로 값을 DTO에 담아서 보낸다. 
+		
+		System.out.println("결과를 말하라" + coupon);
+		
+		// root-contect에서 insert는 regist로 시작으로 지정해놓았다
+		int result = ownerService.registCoupon(coupon);
+		
+		//세션에서 memCode를 담아서 넘겨준다.
+		MemberDTO member = (MemberDTO)session.getAttribute("loginMember");
+		int memCode = member.getMemCode();
+		
+		int result2 = ownerService.registCouponStore(memCode);
+		
+		if(result > 0 && result2 > 0) {
+			ra.addFlashAttribute("message","쿠폰 등록에 성공하였습니다.");
+		} else {
+			ra.addFlashAttribute("message","쿠폰 등록에 실패하였습니다.");
+		}
+		
+		
+		return "redirect:coupon"; // getMapping으로 보내준다
 	}
 	
 	/* 가게정보 수정 */
@@ -69,7 +112,7 @@ public class OwnerController {
 	public void modifyStore(){}
 	
 	@PostMapping("modifyStore")
-	public String updateModifytStore(@RequestParam MultipartFile img, HttpServletRequest request,  Model model) {
+	public String updateModifytStore(@RequestParam MultipartFile img, HttpSession session, HttpServletRequest request, RedirectAttributes ra) {
 		
 		System.out.println("img : " + img);
 		
@@ -81,9 +124,14 @@ public class OwnerController {
 		String phone = request.getParameter("phone");
 		String name = request.getParameter("name");
 		String storeInfo = request.getParameter("storeInfo");
-		String memCode = request.getParameter("memCode");
+		String no = request.getParameter("no");
 		
-		Map<String, String> modifyInfo = new HashMap<>();
+		
+		MemberDTO member = (MemberDTO)session.getAttribute("owner");
+		System.out.println(member);
+		
+		
+		Map<String, Object> modifyInfo = new HashMap<>();
 		modifyInfo.put("storeName", storeName);
 		modifyInfo.put("address", address);
 		modifyInfo.put("dAddress", dAddress);
@@ -92,7 +140,8 @@ public class OwnerController {
 		modifyInfo.put("phone", phone);
 		modifyInfo.put("name", name);
 		modifyInfo.put("storeInfo", storeInfo);
-		modifyInfo.put("memCode",memCode);
+		modifyInfo.put("no",no);
+		modifyInfo.put("member", member);
 		
 		String root = request.getSession().getServletContext().getRealPath("resources");
 		
@@ -103,33 +152,40 @@ public class OwnerController {
 			mkdir.mkdirs();
 		}
 		
-		String orginFileName = img.getOriginalFilename();
-		String ext = orginFileName.substring(orginFileName.indexOf("."));
-		String savedName = UUID.randomUUID().toString().replace("-", "") + ext;
+		if(!img.isEmpty()) {
 		
-		try {
-			img.transferTo(new File(filePath + "/" + savedName));
+			String orginFileName = img.getOriginalFilename();
+			String ext = orginFileName.substring(orginFileName.indexOf("."));
+			String savedName = UUID.randomUUID().toString().replace("-", "") + ext;
 			
-			String fileName = "../resources/uploadFiles/" + savedName;
-			
-			modifyInfo.put("fileName", fileName);
-			
-			int modifyOwner = ownerService.modifyInfo(modifyInfo);
-			int modifyCeo = ownerService.modifyCeo(modifyInfo);
-			int modifyStore = ownerService.modifyStore(modifyInfo);
-			
-			System.out.println("savedName : " + savedName);
-			System.out.println("map : " + modifyInfo);
-			
-			if(modifyOwner > 0 && modifyCeo > 0 && modifyStore > 0) {
+			try {
+				img.transferTo(new File(filePath + "/" + savedName));
 				
-				model.addAttribute("message","변경이 완료되었습니다.");
+				String fileName = "../resources/uploadFiles/" + savedName;
+				
+				modifyInfo.put("fileName", fileName);
+				
+				
+			} catch (IllegalStateException | IOException e) {
+				new File(filePath + "/" + savedName).delete();
+					
+				e.printStackTrace();
 			}
-			
-		} catch (IllegalStateException | IOException e) {
-			new File(filePath + "/" + savedName).delete();
-			model.addAttribute("message","변경에 실패하였습니다.");
-			e.printStackTrace();
+		
+		}
+		
+		int modifyOwner = ownerService.modifyInfo(modifyInfo);
+		int modifyCeo = ownerService.modifyCeo(modifyInfo);
+		int modifyStore = ownerService.modifyStore(modifyInfo);
+		
+		System.out.println("modifyOwner : " + modifyOwner);
+		System.out.println("modifyCeo : " + modifyCeo);
+		System.out.println("modifyStore : " + modifyStore);
+		
+		if(modifyOwner > 0 && modifyCeo > 0 && modifyStore > 0) {
+			ra.addFlashAttribute("message","변경이 완료되었습니다.");
+		} else {
+			ra.addFlashAttribute("message","변경에 실패하였습니다.");
 		}
 		
 		return "redirect:ownerMain";
@@ -139,19 +195,143 @@ public class OwnerController {
 	@GetMapping("modifyOwnerInfo")
 	public void modifyOwnerInfo() {}
 	
-	
-	/* 상품등록 */
-	@GetMapping("productRegist") 
-	public void productRegist(){
+	@PostMapping("modifyOwnerInfo")
+	public ModelAndView modifyOwnerPwd(@RequestParam Map<String, String> info, ModelAndView mv, RedirectAttributes ra,SessionStatus status) {
+		
+		MemberDTO member = new MemberDTO();
+		
+		member.setMemPwd(passwordEncoder.encode(info.get("pwd")));
+		member.setMemCode(Integer.parseInt(info.get("memCode")));
+		
+		int modifyPwd = ownerService.modifyPwd(member);
+		
+		if(modifyPwd > 0) {
+			ra.addFlashAttribute("message","비밀번호가 변경되었습니다.");
+			status.setComplete();
+			mv.setViewName("redirect:/");
+		} else {
+			mv.addObject("message", "비밀번호 변경에 실패하였습니다.");
+			mv.setViewName("modifyOwnerInfo");
+		}
+		
+		return mv;
 		
 	}
 	
+	
+	/* 상품등록 */
+	@GetMapping("productRegist") 
+	public void productRegist(Model model){
+		
+		List<TagDTO> tagList = ownerService.selectTag();
+		List<HashMap<String, String>> categoryList = memberService.selectCategoryList();
+		
+		for(int i = 0; i < categoryList.size(); i++) {
+			
+			System.out.println("키 좀 알려줘 이지삭아 : " + categoryList.get(i).keySet());
+			
+		}
+
+		model.addAttribute("tagList",tagList);
+		model.addAttribute("categoryList",categoryList);
+	}
+	
+	@PostMapping("productRegist")
+	public String registProduct(@RequestParam MultipartFile productImg, HttpSession session, @ModelAttribute ProductDTO product, 
+			HttpServletRequest request, RedirectAttributes rd) {
+		
+		int tag = Integer.parseInt(request.getParameter("tag1"));
+		int tag2 = Integer.parseInt(request.getParameter("tag2"));
+		int tag3 = Integer.parseInt(request.getParameter("tag3"));
+		
+		List<Integer> tagList = new ArrayList<Integer>();
+		tagList.add(tag);
+		tagList.add(tag2);
+		tagList.add(tag3);
+		
+		MemberDTO member = (MemberDTO)session.getAttribute("loginMember");
+		int memCode = member.getMemCode();
+		
+		Map<String,Object> productInfo = new HashMap<String, Object>();
+		productInfo.put("product",product);
+		productInfo.put("tagList", tagList);
+		productInfo.put("memCode", memCode);
+		
+		String root = request.getSession().getServletContext().getRealPath("resources");
+		
+		String filePath = root + "/uploadFiles";
+		
+		File mkdir = new File(filePath);
+		if(!mkdir.exists()) {
+			mkdir.mkdirs();
+		}
+		
+		if(!productImg.isEmpty()) {
+		
+			String orginFileName = productImg.getOriginalFilename();
+			String ext = orginFileName.substring(orginFileName.indexOf("."));
+			String savedName = UUID.randomUUID().toString().replace("-", "") + ext;
+			
+			try {
+				productImg.transferTo(new File(filePath + "/" + savedName));
+				
+				String fileName = "../resources/uploadFiles/" + savedName;
+				
+				productInfo.put("fileName", fileName);
+				
+				
+			} catch (IllegalStateException | IOException e) {
+				new File(filePath + "/" + savedName).delete();
+					
+				e.printStackTrace();
+			}
+		
+		}
+		
+		int result = ownerService.registProduct(productInfo);
+		
+		if(result > 0) {
+			rd.addFlashAttribute("message","상품이 등록되었습니다.");
+		} else {
+			rd.addFlashAttribute("message","상품 등록에 실패하였습니다.");
+		}
+		
+		
+		return "redirect:productManagement";
+	}
+	
+	/* 리뷰 관리 */
+	@GetMapping("review")
+	public String selectReview(@ModelAttribute("loginMember") ReviewDTO review, Model model) {
+		
+		List<ReviewDTO> reviews = ownerService.selectReview(review);
+		System.out.println(reviews);	
+		
+		return "owner/review";	
+		
+	}
+	
+	
 	/* 판매상품 관리 */
 	@GetMapping("productManagement")
-	public void productManagement() {}
+	public void productManagement(Model model, HttpSession session) {
+		
+		MemberDTO member = (MemberDTO)session.getAttribute("loginMember");
+		
+		int memCode = member.getMemCode();
+		
+		List<ProductDTO> productList = ownerService.selectProduct(memCode);
+		
+		model.addAttribute("productList",productList);
+		
+	}
 	
 	
-	/* 리뷰 */
-	@GetMapping("review")
-	public void review() {}
+
+	
+	
+	
+	
+	
+	
 }
